Evaluation refactoring E01.
